Import from Sphinx master
This repo was generated by splitting out the history of the 'text'
writer in its various iterations from the main Sphinx tree [1] at commit
2810f77ed.

This was achieved as follows:

  $ mkdir rst2txt && cd rst2txt && git init && cd -
  $ git clone https://github.com/sphinx-doc/sphinx && cd sphinx
  $ git checkout -b text-writer-split && git checkout -
  $ PATHS_TO_KEEP="sphinx/text sphinx/writers/text.py sphinx/textwriter.py"
  $ git filter-branch -f --index-filter "git rm --ignore-unmatch \
      --cached -qr -- . && git reset -q \$GIT_COMMIT \
      -- $PATHS_TO_KEEP" --prune-empty -- "text-writer-split"

which, in brief, filters out the history for the writer, in the various
locations it has been located. After this, the patch was rebased onto an
empty "initial commit" so we could get rid of the 900 or so merge
commits. This resulted in some merge conflicts and a handful of now
empty commits. These were mostly the result of people duplicating effort
on various branches. Finally, the first commit for this file was edited
to drag the writer out of the 'sphinx/writers' directory and into the
top-level.

The end result is something that does not yet work without Sphinx but
which can be used to give us the basic evolution of the patch.

[1] https://github.com/sphinx-doc/sphinx

Signed-off-by: Stephen Finucane <stephen@that.guru>